Reset: fix
